queries are done for home page
