+++ data-access-layer/utils.js
@@ -1,4 +1,5 @@
 const { User, Passport, Stamp, Country, Tag } = require("../db/models");
+const db = require('../db/models');
 const bcrypt = require("bcryptjs");
 const { Op } = require("sequelize");
 const updateEmail = async (id, email, newEmail) => {
@@ -270,6 +271,96 @@ const getTopCountries = async (user_id) => {
   return arr;
 };
 
+const topWantToVisit = async () => {
+  const topCountries = await Stamp.findAll({
+    where: { rating: 5 },
+    include: [{
+      model: Passport,
+      where: { passport_status: "Want to visit" },
+    },
+    {
+      model: Country
+    }
+    ],
+  });
+  const countryObj = {};
+
+  const countryNames = topCountries.map(data => {
+    return data.Country.dataValues.name
+  });
+
+  countryNames.forEach(country => {
+    if (countryObj[country]) {
+      countryObj[country]++;
+    } else {
+      countryObj[country] = 1;
+    };
+  });
+
+  let countriesArr = Object.entries(countryObj).sort((a, b) => b[1] - a[1]).map(el => el[0])
+  return countriesArr.splice(0, 10);
+}
+
+const topVisited = async () => {
+  const topCountries = await Stamp.findAll({
+    where: { rating: 5 },
+    include: [{
+      model: Passport,
+      where: { passport_status: "visited" },
+    },
+    {
+      model: Country
+    }
+    ],
+  });
+  const countryObj = {};
+
+  const countryNames = topCountries.map(data => {
+    return data.Country.dataValues.name
+  });
+
+  countryNames.forEach(country => {
+    if (countryObj[country]) {
+      countryObj[country]++;
+    } else {
+      countryObj[country] = 1;
+    };
+  });
+
+  let countriesArr = Object.entries(countryObj).sort((a, b) => b[1] - a[1]).map(el => el[0])
+  return countriesArr.splice(0, 10);
+}
+
+const topNearBy = async () => {
+  const topCountries = await Stamp.findAll({
+    where: { rating: 5 },
+    include: [{
+      model: Passport,
+      where: { passport_status: "Near By" },
+    },
+    {
+      model: Country
+    }
+    ],
+  });
+  const countryObj = {};
+
+  const countryNames = topCountries.map(data => {
+    return data.Country.dataValues.name
+  });
+
+  countryNames.forEach(country => {
+    if (countryObj[country]) {
+      countryObj[country]++;
+    } else {
+      countryObj[country] = 1;
+    };
+  });
+
+  let countriesArr = Object.entries(countryObj).sort((a, b) => b[1] - a[1]).map(el => el[0])
+  return countriesArr.splice(0, 10);
+}
+
 module.exports = {
   get100Stamps,
   updateEmail,
@@ -290,4 +381,7 @@ module.exports = {
   getComments,
   getTags,
   getTopCountries,
+  topWantToVisit,
+  topVisited,
+  topNearBy,
 };
